[refactor] Moved 'compareArrays' to its own TS file
