+++ src/Script.js
@@ -36,7 +36,7 @@ import {parse, Node}                            from "../utils/acorn";
 import {dialogBoxCreate}                        from "../utils/DialogBox";
 import {Reviver, Generic_toJSON,
         Generic_fromJSON}                       from "../utils/JSONReviver";
-import {compareArrays}                          from "../utils/HelperFunctions";
+import {compareArrays}                          from "../utils/helpers/compareArrays";
 import {createElement}                          from "../utils/uiHelpers/createElement";
 import {formatNumber}                           from "../utils/StringHelperFunctions";
 import {roundToTwo}                             from "../utils/helpers/roundToTwo";
